{{cookiecutter.project_name}}/requirements/local.txt
-r base.txt

# Testing
# ------------------------------------------------------------------------------
pytest
pytest-django
factory-boy
faker
coverage

# Code quality
# ------------------------------------------------------------------------------
flake8 # https://github.com/PyCQA/flake8